Ещё один небольшой рефактор с перекидыванием функций
