+++ python_interface/handlers_functions/point_estimate/estimation.py
@@ -1,6 +1,6 @@
 from rpy2 import robjects as r_obj
-from handlers_functions.standard_functions.standart_functions import (list_is_integer as is_int,
-                                                                      mean, var_unbased as uvar)
+from handlers_functions.standard_functions.math_functions import mean, var_unbased as uvar
+from handlers_functions.standard_functions.standart_functions import list_is_integer as is_int
 # value estimation means finding mean() from the data || оценка значения равняется поиску выборочного среднего
 r_sum = r_obj.r["sum"]  # defining global R function for processing likelihood ||
                         # определение функции R для определения правдоподобности
